Move Inner Factory classes to outside for improved handling on older SDK.
